# fix: always return uint8arrays