--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -43,7 +43,16 @@ jobs:
       - uses: actions/checkout@v2
       - uses: microsoft/playwright-github-action@v1
       - run: npm install
-      - run: npx aegir test -t browser -t webworker --bail
+      - run: npx aegir test -t browser --bail --cov
+      - uses: codecov/codecov-action@v1
+  test-chrome-webworker:
+    needs: check
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: microsoft/playwright-github-action@v1
+      - run: npm install
+      - run: npx aegir test -t webworker --bail
   test-firefox:
     needs: check
     runs-on: ubuntu-latest
@@ -51,7 +60,15 @@ jobs:
       - uses: actions/checkout@v2
       - uses: microsoft/playwright-github-action@v1
       - run: npm install
-      - run: npx aegir test -t browser -t webworker --bail -- --browser firefox
+      - run: npx aegir test -t browser --bail -- --browser firefox
+  test-firefox-webworker:
+    needs: check
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: microsoft/playwright-github-action@v1
+      - run: npm install
+      - run: npx aegir test -t webworker --bail -- --browser firefox
   test-webkit:
     needs: check
     runs-on: ubuntu-latest
@@ -59,7 +76,15 @@ jobs:
       - uses: actions/checkout@v2
       - uses: microsoft/playwright-github-action@v1
       - run: npm install
-      - run: npx aegir test -t browser -t webworker --bail -- --browser webkit
+      - run: npx aegir test -t browser --bail -- --browser webkit
+  test-webkit-webworker:
+    needs: check
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: microsoft/playwright-github-action@v1
+      - run: npm install
+      - run: npx aegir test -t webworker --bail -- --browser webkit
   test-electron-main:
     needs: check
     runs-on: ubuntu-latest