--- a/benchmarks/alloc.js
+++ b/benchmarks/alloc.js
@@ -1,8 +1,8 @@
 /* eslint-disable no-console */
 
 /*
-$ node benchmarks/to-string.js
-$ npx playwright-test benchmarks/to-string.js --runner benchmark
+$ node benchmarks/alloc.js
+$ npx playwright-test benchmarks/alloc.js --runner benchmark
 */
 
 import Benchmark from 'benchmark'