--- a/src/from-string.js
+++ b/src/from-string.js
@@ -1,4 +1,5 @@
 import bases from './util/bases.js'
+import { asUint8Array } from './util/as-uint8array.js'
 
 /**
  * @typedef {import('./util/bases').SupportedEncodings} SupportedEncodings
@@ -23,7 +24,7 @@ export function fromString (string, encoding = 'utf8') {
   }
 
   if ((encoding === 'utf8' || encoding === 'utf-8') && globalThis.Buffer != null && globalThis.Buffer.from != null) {
-    return globalThis.Buffer.from(string, 'utf8')
+    return asUint8Array(globalThis.Buffer.from(string, 'utf-8'))
   }
 
   // add multibase prefix