--- a/src/xor.js
+++ b/src/xor.js
@@ -1,4 +1,5 @@
 import { allocUnsafe } from './alloc.js'
+import { asUint8Array } from './util/as-uint8array.js'
 
 /**
  * Returns the xor distance between two arrays
@@ -17,5 +18,5 @@ export function xor (a, b) {
     result[i] = a[i] ^ b[i]
   }
 
-  return result
+  return asUint8Array(result)
 }