--- a/test/xor.spec.js
+++ b/test/xor.spec.js
@@ -24,4 +24,14 @@ describe('Uint8Array xor', () => {
 
     expect(xor(a, b)).to.deep.equal(Uint8Array.from([0, 0]))
   })
+
+  it('xors returns Uint8Array', () => {
+    const a = Uint8Array.from([1, 1])
+    const b = Uint8Array.from([1, 1])
+    const c = xor(a, b)
+    const slice = c.slice()
+
+    // node slice is a copy operation, Uint8Array slice is a no-copy operation
+    expect(slice.buffer).to.not.equal(c.buffer)
+  })
 })